==> notebook/1_tokenize.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# Prepare Data

CHARS = [
  '\n', ' ', '!', '"', '#', '$', '%', '&', "'", '(', ')', '*', '+',
  ',', '-', '.', '/', '0', '1', '2', '3', '4', '5', '6', '7', '8',
  '9', ':', ';', '<', '=', '>', '?', '@', 'A', 'B', 'C', 'D', 'E',
  'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R',
  'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z', '[', '\\', ']', '^', '_',
  'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm',
  'n', 'o', 'other', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y',
  'z', '}', '~', 'ก', 'ข', 'ฃ', 'ค', 'ฅ', 'ฆ', 'ง', 'จ', 'ฉ', 'ช',
  'ซ', 'ฌ', 'ญ', 'ฎ', 'ฏ', 'ฐ', 'ฑ', 'ฒ', 'ณ', 'ด', 'ต', 'ถ', 'ท',
  'ธ', 'น', 'บ', 'ป', 'ผ', 'ฝ', 'พ', 'ฟ', 'ภ', 'ม', 'ย', 'ร', 'ฤ',
  'ล', 'ว', 'ศ', 'ษ', 'ส', 'ห', 'ฬ', 'อ', 'ฮ', 'ฯ', 'ะ', 'ั', 'า',
  'ำ', 'ิ', 'ี', 'ึ', 'ื', 'ุ', 'ู', 'ฺ', 'เ', 'แ', 'โ', 'ใ', 'ไ',
  'ๅ', 'ๆ', '็', '่', '้', '๊', '๋', '์', 'ํ', '๐', '๑', '๒', '๓',
  '๔', '๕', '๖', '๗', '๘', '๙', '‘', '’', '\ufeff'
]
CHARS_MAP = {v: k for k, v in enumerate(CHARS)}

def create_df_with_context(df, n_pad):
    for i in range(n_pad):
        df['char-{}'.format(i+1)] = df['char'].shift(i + 1)
        df['char{}'.format(i+1)] = df['char'].shift(-i - 1)
    return df[n_pad: -n_pad]

def prepare_best_data(option = 'train', n_pad = 10):
    pad = [{'char': ' ', 'target': True}]
    df_pad = pd.DataFrame(pad * n_pad)
    
    df_list = []
    article_types = ['article', 'encyclopedia', 'news', 'novel']
    for article_type in article_types:
        df_list.append(pd.read_csv(os.path.join('../../corpora/BEST', option, 'df_best_{}_{}.csv'.format(article_type, option))))
  
    df_list = [df_pad] + df_list + [df_pad]
    df = pd.concat(df_list)

    df['char'] = df['char'].map(lambda x: CHARS_MAP.get(x, CHARS_MAP['other']))
    
    df_with_context = create_df_with_context(df, n_pad)

    char_row = ['char' + str(i + 1) for i in range(n_pad)] + ['char-' + str(i + 1) for i in range(n_pad)] + ['char']

    x_char = df_with_context[char_row].as_matrix()
    y = df_with_context['target'].astype(int).as_matrix()

    return x_char, y

x_train_char, y_train = prepare_best_data('train', 10)
x_val_char, y_val = prepare_best_data('val', 10)
x_test_char, y_test = prepare_best_data('test', 10)

print('Training data shape          :', x_train_char.shape)
print('Training data labels shape   :', y_train.shape)
print('Validation data shape        :', x_val_char.shape)
print('Validation data labels shape :', y_val.shape)
print('Test data shape              :', x_test_char.shape)
print('Test data labels shape       :', y_test.shape)

# Print some entry from the data to make sure it is the same as what you think.
print('First 3 features: ', x_train_char[:3])
print('First 30 class labels', y_train[:30])

# Convolution Neural Network

from keras.models import Model
from keras.layers import Input, Conv1D, Embedding, TimeDistributed, Flatten, Dense
from keras.optimizers import Adam

def get_cnn():
    input1 = Input(shape=(21,))
    x = Embedding(len(CHARS), 32)(input1)
    x = Conv1D(100, 5, strides=1, activation='relu', padding="same")(x)
    x = TimeDistributed(Dense(5))(x)
    x = Flatten()(x)
    x = Dense(100, activation='relu')(x)
    out = Dense(1, activation='sigmoid')(x)
    model = Model(inputs=input1, outputs=out)
    model.compile(optimizer=Adam(), loss='binary_crossentropy', metrics=['acc'])
    return model

from sklearn.metrics import f1_score,precision_score,recall_score

def evaluate(x_test, y_test, model):
    y_pred = model.predict(x_test)
    prob_to_class = lambda p: 1 if p[0]>=0.5 else 0
    y_pred = np.apply_along_axis(prob_to_class,1,y_pred)
    f1score = f1_score(y_test,y_pred)
    precision = precision_score(y_test,y_pred)
    recall = recall_score(y_test,y_pred)
    return f1score, precision, recall

from keras import backend as K
K.clear_session()

print('start training')
verbose = 1
model_cnn = get_cnn()

train_params = [(3, 512)]
for (epochs, batch_size) in train_params:
    print("train with {} epochs and {} batch size".format(epochs, batch_size))
    model_cnn.fit(x_train_char, y_train, epochs=epochs, batch_size=batch_size, verbose=verbose,
                           validation_data=(x_val_char, y_val))

evaluate(x_test_char, y_test, model_cnn)

path_model_cnn='../data/model_tokenize_cnn.h5'
model_cnn.save(path_model_cnn)

# Load from saved model

from keras.models import load_model
path_model_cnn='../data/model_tokenize_cnn.h5'
model_cnn = load_model(path_model_cnn)

# Tokenize Wongnai Data

tr = pd.read_csv("../data/w_review_train.csv", sep=";", names = ["review", "label"])
ts = pd.read_csv("../data/test_file.csv", sep=";")

def prepare_feature(text, n_pad = 10):
    pad = [{'char': ' '}]
    df_pad = pd.DataFrame(pad * n_pad)
    
    df_list = []
    df_list.append(pd.DataFrame({'char': list(text)}))
    df_list = [df_pad] + df_list + [df_pad]
                   
    df = pd.concat(df_list)

    df['char'] = df['char'].map(lambda x: CHARS_MAP.get(x, CHARS_MAP['other']))
    
    df_with_context = create_df_with_context(df, n_pad)

    char_row = ['char' + str(i + 1) for i in range(n_pad)] + ['char-' + str(i + 1) for i in range(n_pad)] + ['char']

    x_char = df_with_context[char_row].as_matrix()

    return x_char

def char_to_word(text, y_pred):
    split_text=""
    for char, y in zip(text, y_pred):
        if y == 1:
            split_text+=" "
            split_text+=char
        else:
            split_text+=char
    return split_text.split(" ")

def tokenize(text, model):
    x_char = prepare_feature(text)
    y_pred = model.predict(x_char)
    prob_to_class = lambda p: 1 if p[0]>=0.5 else 0
    y_pred = np.apply_along_axis(prob_to_class,1,y_pred)
    tokens = char_to_word(text, y_pred)
    return tokens

# This take around an hour to finish
tr["tokenizedReview"] = tr["review"].apply(lambda x: tokenize(x,model_cnn))

tr.head()

ts["tokenizedReview"] = ts["review"].apply(lambda x: tokenize(x,model_cnn))

ts.head()

tr.to_csv('../data/tokenized_train.csv', index=False)
ts.to_csv('../data/tokenized_test.csv', index=False)
