==> notebook/2_feature_extraction.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# Load Tokenized Data

tr = pd.read_csv("../data/tokenized_train.csv")
ts = pd.read_csv("../data/tokenized_test.csv")

def preprocess(tokens):
    tokens = eval(tokens)
    tokens = list(filter(lambda x: x!='', tokens))
    tokens = list(map(lambda x: x.lower(), tokens))
    return tokens

tr["tokenizedReview"] = tr["tokenizedReview"].apply(lambda x: preprocess(x))
ts["tokenizedReview"] = ts["tokenizedReview"].apply(lambda x: preprocess(x))

tr.head()

ts.head()

# TF-IDF

def tokenize(word_list):
    return word_list

from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(tokenizer=tokenize, lowercase=False)
trainTfIdf = vectorizer.fit_transform(tr["tokenizedReview"])
testTfIdf = vectorizer.transform(ts["tokenizedReview"])

trainTfIdf.shape

testTfIdf.shape

# SVD 

from sklearn.decomposition import TruncatedSVD

svd = TruncatedSVD(n_components=300, n_iter=10, random_state=42)
trainFeatures = svd.fit_transform(trainTfIdf)
testFeatures = svd.transform(testTfIdf)

for i in range(300):
    tr["feature"+str(i)] = trainFeatures[:,i]
    ts["feature"+str(i)] = testFeatures[:,i]

tr.drop("tokenizedReview", axis=1, inplace=True)
ts.drop("tokenizedReview", axis=1, inplace=True)

tr.head()

ts.head()

tr.to_csv('../data/train_with_feature.csv', index=False)
ts.to_csv('../data/test_with_feature.csv', index=False)
