==> notebook/3_rating_prediction.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
%matplotlib inline

# Prepare Data

tr = pd.read_csv("../data/train_with_feature.csv")
ts = pd.read_csv("../data/test_with_feature.csv")

tr.head(3)

ts.head(3)

split = int(tr.shape[0]*9/10)
features = ["feature"+str(i) for i in range(300)]
x_train = tr[features][:split].as_matrix()
y_train = tr["label"][:split].as_matrix()
y_train_one_hot = pd.get_dummies(tr["label"][:split]).as_matrix()
x_val = tr[features][split:].as_matrix()
y_val = tr["label"][split:].as_matrix()
y_val_one_hot = pd.get_dummies(tr["label"][split:]).as_matrix()
x_test = ts[features].as_matrix()

print('Training data shape                  :', x_train.shape)
print('Training data labels shape           :', y_train.shape)
print('Training data labels one hot shape   :', y_train_one_hot.shape)
print('Validation data shape                :', x_val.shape)
print('Validation data labels shape         :', y_val.shape)
print('Validation data labels one hot shape :', y_val_one_hot.shape)
print('Test data shape                      :', x_test.shape)

# Feedforward Nueral Network - Regression

from keras.models import Model
from keras.layers import Input, Dense, Dropout
from keras.optimizers import Adam

def get_nn_regression():
    input1 = Input(shape=(300,))
    x = Dense(100, activation='relu')(input1)
    x = Dropout(0.5)(x)
    x = Dense(10, activation='relu')(x)
    x = Dropout(0.1)(x)
    out = Dense(1, activation='relu')(x)
    model = Model(inputs=input1, outputs=out)
    model.compile(optimizer=Adam(), loss='mean_squared_error', metrics=['acc'])
    return model

from keras import backend as K
K.clear_session()

print('start training')
verbose = 1
model_nn_regression = get_nn_regression()

train_params = [(20, 32)]
for (epochs, batch_size) in train_params:
    print("train with {} epochs and {} batch size".format(epochs, batch_size))
    model_nn_regression.fit(x_train, y_train, epochs=epochs, batch_size=batch_size, verbose=verbose,
                           validation_data=(x_val, y_val))

#### Calculate F1 score on Validation Data

from sklearn.metrics import f1_score
y_pred_val = model_nn_regression.predict(x_val)
y_pred_val = np.rint(y_pred_val).astype(int)
f1_score(y_val, y_pred_val,average='micro')

#### Predict Rating For Test Data

y_pred = model_nn_regression.predict(x_test)
y_pred = np.rint(y_pred).astype(int)
y_pred.shape

pred = ts[["reviewID"]].copy()
pred["rating"] = y_pred

pred.head(10)

pred.to_csv("../data/prediction_nn_regression.csv", index=False)

# Feedforward Nueral Network - Classification

from keras.models import Model
from keras.layers import Input, Dense, Dropout
from keras.optimizers import Adam

def get_nn_classification():
    input1 = Input(shape=(300,))
    x = Dense(100, activation='relu')(input1)
    x = Dropout(0.5)(x)
    x = Dense(10, activation='relu')(x)
    x = Dropout(0.1)(x)
    out = Dense(5, activation='softmax')(x)
    model = Model(inputs=input1, outputs=out)
    model.compile(optimizer=Adam(), loss='categorical_crossentropy', metrics=['acc'])
    return model

from keras import backend as K
K.clear_session()

print('start training')
verbose = 1
model_nn_classification = get_nn_classification()

train_params = [(20, 32)]
for (epochs, batch_size) in train_params:
    print("train with {} epochs and {} batch size".format(epochs, batch_size))
    model_nn_classification.fit(x_train, y_train_one_hot, epochs=epochs, batch_size=batch_size, verbose=verbose,
                           validation_data=(x_val, y_val_one_hot))

#### Calculate F1 score on Validation Data

from sklearn.metrics import f1_score
y_pred_distribution_val = model_nn_classification.predict(x_val)
y_pred_val = y_pred_distribution_val.argmax(axis=1)+1
f1_score(y_pred_val, y_val,average='micro')

#### Predict Rating For Test Data

y_pred_distribution = model_nn_classification.predict(x_test)
y_pred = y_pred_distribution.argmax(axis=1)+1
y_pred.shape

pred = ts[["reviewID"]].copy()
pred["rating"] = y_pred

pred.head(10)

pred.to_csv("../data/prediction_nn_classification.csv", index=False)
